requirements.txt
feedparser>=6.0.0
pyyaml>=6.0
feedgen>=0.9.0
requests>=2.25.0
arxiv>=1.4.0
apscheduler>=3.7.0
flask>=2.0.0
flask-cors>=3.0.10
nltk